// management/src/routes/comments/__generated__/CommentsTableFragment.graphql.ts
export type CommentStatus = "approved" | "pending" | "spam" | "trash" | "%future added value";